requirements.txt
Django>=4.2.16,<4.3
djangorestframework>=3.15.2,<3.16
flake8
psycopg2>=2.8.6,<2.9
drf-spectacular
pillow